API Reference

Core types

ADDM.Trial — Type

Trial(choice, RT, valueLeft, valueRight)

Arguments

Required keyword arguments

choice: either -1 (for left item) or +1 (for right item).

RT: response time in milliseconds.

valueLeft: value of the left item.

valueRight: value of the right item.

Optional

RDV: vector of RDV over time.

fixItem: list of items fixated during the trial in chronological order; 1 correponds to left, 2 corresponds to right, and any other value is considered a transition/blank fixation.

fixTime: list of fixation durations (in milliseconds) in chronological order.

fixRDV: list of Float64 corresponding to the RDV values at the end of each fixation in the trial.

uninterruptedLastFixTime: Int64 corresponding to the duration, in milliseconds, that the last fixation in the trial would have if it had not been interrupted when a decision was made.

Example

julia> t = Trial(choice = 1, RT = 2145, valueLeft = 1, valueRight = 3)
Trial(1, 2145, 1, 3, #undef, #undef, #undef, #undef, #undef)

julia> t.RT
2145

julia> t.uninterruptedLastFixTime
ERROR: UndefRefError: access to undefined reference
Stacktrace:
 [1] getproperty(x::Trial, f::Symbol)
   @ Base ./Base.jl:37
 [2] top-level scope
   @ REPL[4]:1

julia> t.uninterruptedLastFixTime = 189
189

julia> t
Trial(1, 2145, 1, 3, #undef, #undef, #undef, 189, #undef)

ADDM.aDDM — Type

Constructor for model definitions that will contain model parameter and parameter value mapping. Not intended to be used alone but as part of define_model

Example

julia> MyModel = ADDM.aDDM()
aDDM(Dict{Symbol, Any}())

julia> MyModel.d = 0.005
0.005

julia> MyModel.σ = .06
0.06

julia> MyModel
aDDM(Dict{Symbol, Any}(:σ => 0.06, :d => 0.005))

ADDM.define_model — Function

define_model(d, σ, θ = 1, η = 0, barrier = 1, decay = 0, nonDecisionTime = 0, bias = 0.0)

Create attentional drift diffusion model with parameters described in Krajbich et al. (2010).

Arguments

Required parameters

d: Number, parameter of the model which controls the speed of integration of the signal.

σ: Number, parameter of the model, standard deviation for the normal distribution.

Optional parameters

θ: Float64 Traditionally between 0 and 1, parameter of the model which controls the attentional discounting. Default at 1 makes it a ddm.

η: Float64 Additive attentional enhancement the attentional discounting. Default at 0 makes it a ddm.

barrier: positive Int64, boundary separation in each direction from 0. Default at 1.

decay: constant for linear barrier decay at each time step. Default at 0.

nonDecisionTime: non-negative Number, the amount of time in milliseconds during which processes other than evidence accummulation occurs. Default at 0.

bias: Number, corresponds to the initial value of the relative decision value variable. Must be smaller than barrier.

Todo

Tests

Change decay parameter to function instead of scalar

Example

julia julia> MyModel = define_model(d = .006, σ = 0.05) aDDM(Dict{Symbol, Any}(:nonDecisionTime => 0, :σ => 0.05, :d => 0.006, :bias => 0.0, :barrier => 1, :decay => 0, :θ => 1.0, :η => 0.0))`

Fixation data

ADDM.FixationData — Type

FixationData(probFixLeftFirst, latencies, transitions, fixations; 
             fixDistType="fixation")

Arguments:

probFixLeftFirst: Float64 between 0 and 1, empirical probability that the left item will be fixated first.

latencies: Vector corresponding to the empirical distribution of trial latencies (delay before first fixation) in milliseconds.

transitions: Vector corresponding to the empirical distribution of transitions (delays between item fixations) in milliseconds.

fixations: Dict whose indexing is defined according to parameter fixDistType. Each entry is an array corresponding to the empirical distribution of item fixation durations in milliseconds.

fixDistType: String, one of {'simple', 'difficulty', 'fixation'}, determines how the fixation distributions are indexed. If 'simple', fixation distributions are indexed only by type (1st, 2nd, etc). If 'difficulty', they are indexed by type and by trial difficulty, i.e., the absolute value for the trial's value difference. If 'fixation', they are indexed by type and by the value difference between the fixated and unfixated items.

ADDM.process_fixations — Function

process_fixations(data::Dict; timeStep::Number = 10, 
                                 maxFixTime::Number = 3000, 
                                 numFixDists::Int64 = 3, fixDistType::String = "fixation", 
                                 valueDiffs::Vector{Int64} = collect(-3:1:3), 
                                 subjectIds::Vector{String} = String[])

Create empirical distributions from the data to be used when generating model simulations.

Arguments

data: a dict, indexed by subjectId, where each entry is a list of Trial objects. E.g. output of load_data_from_csv

timeStep: integer, minimum duration of a fixation to be considered, in miliseconds.

maxFixTime: integer, maximum duration of a fixation to be considered, in miliseconds.

numFixDists: integer, number of fixation types to use in the fixation distributions. For instance, if numFixDists equals 3, then 3 separate fixation types will be used, corresponding to the 1st, 2nd and other (3rd and up) fixations in each trial.

fixDistType: string, one of {'simple', 'difficulty', 'fixation'}, used to determine how the fixation distributions should be indexed. If 'simple', then fixation distributions will be indexed only by type (1st, 2nd, etc). If 'difficulty', they will be indexed by type and by trial difficulty, i.e., the absolute value for the trial's value difference. If 'fixation', they will be indexed by type and by the value difference between the fixated and unfixated items.

valueDiffs: list of integers. If fixDistType is 'difficulty' or 'fixation', valueDiffs is a range correspoding to the item values to be used when indexing the fixation distributions.

subjectIds: list of strings corresponding to the subjects whose data should be used. If not provided, all existing subjects will be used.

Return

A FixationData object.

ADDM.convert_to_fixationDist — Function

convert_to_fixationDist(fixationData::FixationData; timeStep::Number = 10)

Create empirical distributions from the data to be used when generating model simulations.

Arguments

fixationData: FixationData type that is the output of process_fixations.

timeStep: integer, timeBin size in ms.

Return

fixationDist: Dictionary indexed by value difference and fixation type. Contains the distributions of fixation durations in each time bin specifiu

timeBinMidPoints: Mid points of the time bins, for which the fixation duration distributions were calculated. Will be the durations sampled in addm_simulate_trial if using fixationDist instead of fixationData

Data simulation

ADDM.aDDM_simulate_trial — Function

aDDM_simulate_trial(model::aDDM, fixationData::FixationData, 
                    valueLeft::Number, valueRight::Number; timeStep::Number=10.0, 
                    numFixDists::Int64=3 , fixationDist=nothing, timeBins=nothing, 
                    cutOff::Number=100000)

Generate a DDM trial given the item values.

Arguments

model: aDDM object.

fixationData: FixationData object. Required even when using fixationDist because it specifies latencies and transitions as well.

valueLeft: value of the left item.

valueRight: value of the right item.

timeStep: Number, value in milliseconds to be used for binning time axis.

numFixDists: Int64, number of fixation types to use in the fixation distributions. For instance, if numFixDists equals 3, then 3 separate fixation types will be used, corresponding to the 1st, 2nd and other (3rd and up) fixations in each trial.

fixationDist: distribution of fixations which, when provided, will be used instead of fixationData.fixations. This should be a dict of dicts of dicts, corresponding to the probability distributions of fixation durations. Indexed first by fixation type (1st, 2nd, etc), then by the value difference between the fixated and unfixated items, then by time bin. Each entry is a number between 0 and 1 corresponding to the probability assigned to the particular time bin (i.e. given a particular fixation type and value difference, probabilities for all bins should add up to 1). Can be obtained from fixationData using convert_to_fixationDist. If using this instead of fixationData to sample fixations make sure to specify latency and transition info in fixationData.

timeBins: array containing the time bins used in fixationDist. Can be obtained fromfixationData using convert_to_fixationDist

Returns

An Trial object resulting from the simulation.

ADDM.DDM_simulate_trial — Function

DDM_simulate_trial(model::aDDM, valueLeft::Number, valueRight::Number; timeStep::Number = 10.0, 
                   cutOff::Int64 = 20000)

Generate a DDM trial given the item values.

Arguments

model: aDDM object.

valueLeft: value of the left item.

valueRight: value of the right item.

timeStep: Number, value in milliseconds to be used for binning the time axis.

cutOff: Number, value in milliseconds to be used as a cap if trial response time is too long.

Returns

A Trial object resulting from the simulation.

ADDM.simulate_data — Function

simulate_data(model::aDDM, stimuli, simulator_fn, simulator_args = (timeStep = 10.0, cutOff = 20000))

Simulate data using the model for the given stimuli.

Arguments

model: aDDM object.

stimuli: Named tuple with valueLeft and valueRight specifying the values of options.

simulator_fn: Name of the function that simulates a trial for the given model.

simulator_args: Named tuple containing kwargs that should be fed to simulator_fn

Returns

Vector of Trial objects containing simulated data.

Likelihood computation

ADDM.aDDM_get_trial_likelihood — Function

aDDM_get_trial_likelihood(addm::aDDM, trial::aDDMTrial; timeStep::Number = 10.0, 
                          approxStateStep::Number = 0.1)

Compute the likelihood of the data from a single aDDM trial for these particular aDDM parameters.

Arguments:

model: aDDM object.

trial: Trial object.

timeStep: Number, value in milliseconds to be used for binning the time axis.

approxStateStep: Number, to be used for binning the RDV axis.

Returns:

The likelihood obtained for the given trial and model.

ADDM.DDM_get_trial_likelihood — Function

DDM_get_trial_likelihood(ddm::aDDM, trial::Trial; timeStep::Number = 10, 
                         approxStateStep::Number = 0.1, 
                         decay::Number = 0)

Compute the likelihood of the data from a single DDM trial for these particular DDM parameters.

Arguments

model: aDDM object.

trial: Trial object.

timeStep: Number, value in milliseconds to be used for binning the time axis.

approxStateStep: Number, to be used for binning the RDV axis.

Returns

The likelihood obtained for the given trial and model.

ADDM.compute_trials_nll — Function

compute_trials_nll(model::aDDM, data, likelihood_fn, likelihood_args = (timeStep = 10.0, cutOff = 20000))

Compute likelihood of a dataset for a given model.

Arguments

model: aDDM object.

data: Vector of ADDM.Trial objects.

likelihood_fn: Name of the function that computes the likelhoods of a trial for the given model.

likelihood_args: Named tuple containing kwargs that should be fed to likelihood_fn

return_trial_likelihoods: Boolean to specify whether to return the likelihoods for each trial

Returns

Negative log likelihood of data

Grid search

ADDM.grid_search — Function

grid_search(data, param_grid, likelihood_fn, return_grid_likelihoods = false; 
            likelihood_args =  (timeStep = 10.0, approxStateStep = 0.1), return_trial_likelihoods = false)

Compute the likelihood of either observed or simulated data for all parameter combinations in paramGrid.

Arguments

data:

Returns:

...

Helpers

ADDM.load_data_from_csv — Function

load_data_from_csv(expdataFileName, fixationsFileName; convertItemValues=nothing)

Load experimental data from two CSV files: an experimental data file and a fixations file. Format expected for experimental data file: parcode, trial, rt, choice, itemleft, itemright. Format expected for fixations file: parcode, trial, fixitem, fixtime.

Arguments

expdataFileName: String, name of experimental data file.

fixationsFileName: String, name of fixations file.

parcode: Subject identifier

trial: Trial number

fix_item: Fixation location. 0 = transition, 1 = left, 2 = right,

3 = latency.

fix_time: Fixation duration.

Return

A dict, indexed by subjectId, where each entry is a list of Trial objects.

ADDM.convert_param_symbols — Function

convert_param_symbols(model)

Convert parameter names that are specified in text into greek/latex symbols. Used by ADDM.grid_search